requirements.txt
fastapi
uvicorn
pytesseract
Pillow
opencv-python-headless
numpy
jinja2
python-multipart
